+++ app/src/main/res/layout/activity_crear_evento.xml
@@ -157,6 +157,23 @@
                             app:layout_constraintHorizontal_bias="0.511"
                             app:layout_constraintStart_toStartOf="parent" />
 
+                        <LinearLayout
+                            android:id="@+id/artistaFormLayout"
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content"
+                            android:orientation="vertical" />
+
+                        <Button
+                            android:id="@+id/addArtistBtn"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:layout_marginStart="30dp"
+                            android:layout_marginEnd="30dp"
+                            android:layout_marginTop="16dp"
+                            android:layout_marginBottom="16dp"
+                            android:backgroundTint="@color/purple_500"
+                            android:text="Añadir Artista" />
+
                         <EditText
                             android:id="@+id/UrlImagenEvento"
                             android:layout_width="wrap_content"
@@ -222,39 +239,52 @@
                                 app:layout_constraintHorizontal_bias="0.511"
                                 app:layout_constraintStart_toStartOf="parent" />
 
-                            <EditText
-                                android:id="@+id/Hora_Inicio_Funcion_1"
+                            <TextView
+                                android:id="@+id/TextViewHoraInicio"
                                 android:layout_width="wrap_content"
                                 android:layout_height="wrap_content"
-                                android:layout_marginStart="30dp"
-                                android:layout_marginTop="28dp"
-                                android:backgroundTint="@color/white"
-                                android:ems="15"
-                                android:hint="Hora de inicio"
-                                android:minHeight="48dp"
+                                android:layout_gravity="center"
+                                android:layout_marginTop="16dp"
+                                android:text="Hora de inicio"
                                 android:textColor="@color/white"
-                                android:textColorHint="@color/white"
-                                app:layout_constraintEnd_toEndOf="parent"
-                                app:layout_constraintHorizontal_bias="0.511"
-                                app:layout_constraintStart_toStartOf="parent" />
+                                android:textSize="20sp"
+                                tools:ignore="MissingConstraints"
+                                tools:layout_editor_absoluteX="23dp"
+                                tools:layout_editor_absoluteY="179dp" />
 
-                            <EditText
-                                android:id="@+id/Hora_Fin_Funcion_1"
+                            <TimePicker
+                                android:id="@+id/timePickerInicio"
                                 android:layout_width="wrap_content"
                                 android:layout_height="wrap_content"
-                                android:layout_marginStart="30dp"
-                                android:layout_marginTop="28dp"
-                                android:layout_marginBottom="16dp"
-                                android:backgroundTint="@color/white"
-                                android:ems="15"
-                                android:hint="Hora de fin"
-                                android:minHeight="48dp"
-                                android:textColor="@color/white"
-                                android:textColorHint="@color/white"
-                                app:layout_constraintEnd_toEndOf="parent"
-                                app:layout_constraintHorizontal_bias="0.511"
-                                app:layout_constraintStart_toStartOf="parent" />
+                                android:layout_centerHorizontal="true"
+                                android:layout_gravity="center"
+                                android:scaleX="0.9"
+                                android:scaleY="0.9"
+                                android:theme="@style/TimePickerTheme"
+                                android:timePickerMode="spinner" />
 
+                            <TextView
+                                android:id="@+id/TextViewHoraFin"
+                                android:layout_width="wrap_content"
+                                android:layout_height="wrap_content"
+                                android:layout_gravity="center"
+                                android:text="Hora de fin"
+                                android:textColor="@color/white"
+                                android:textSize="20sp"
+                                tools:ignore="MissingConstraints"
+                                tools:layout_editor_absoluteX="23dp"
+                                tools:layout_editor_absoluteY="179dp" />
+
+                            <TimePicker
+                                android:id="@+id/timePickerFin"
+                                android:layout_width="wrap_content"
+                                android:layout_height="wrap_content"
+                                android:layout_centerHorizontal="true"
+                                android:layout_gravity="center"
+                                android:scaleX="0.9"
+                                android:scaleY="0.9"
+                                android:theme="@style/TimePickerTheme"
+                                android:timePickerMode="spinner" />
                             <LinearLayout
                                 android:layout_width="match_parent"
                                 android:layout_height="wrap_content"
@@ -272,7 +302,7 @@
                             android:orientation="vertical" />
 
                         <Button
-                            android:id="@+id/addBtn"
+                            android:id="@+id/addFunctionBtn"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="30dp"
